\name{data_table}
\alias{data_table}
\docType{data}
\title{
Data used to reproduce Table 5 results in Jiang et. al. (2022)
}
\description{
Data used to reproduce Table 5 results in Jiang et. al. (2022).
}
\usage{data("data_table")}
\format{
  A data frame with 2159 observations on the following 69 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{X9}}{a numeric vector}
    \item{\code{X10}}{a numeric vector}
    \item{\code{X11}}{a numeric vector}
    \item{\code{X12}}{a numeric vector}
    \item{\code{X13}}{a numeric vector}
    \item{\code{X14}}{a numeric vector}
    \item{\code{X15}}{a numeric vector}
    \item{\code{X16}}{a numeric vector}
    \item{\code{X17}}{a numeric vector}
    \item{\code{X18}}{a numeric vector}
    \item{\code{X19}}{a numeric vector}
    \item{\code{X20}}{a numeric vector}
    \item{\code{X21}}{a numeric vector}
    \item{\code{X22}}{a numeric vector}
    \item{\code{X23}}{a numeric vector}
    \item{\code{X24}}{a numeric vector}
    \item{\code{X25}}{a numeric vector}
    \item{\code{X26}}{a numeric vector}
    \item{\code{X27}}{a numeric vector}
    \item{\code{X28}}{a numeric vector}
    \item{\code{X29}}{a numeric vector}
    \item{\code{X30}}{a numeric vector}
    \item{\code{X31}}{a numeric vector}
    \item{\code{X32}}{a numeric vector}
    \item{\code{X33}}{a numeric vector}
    \item{\code{X34}}{a numeric vector}
    \item{\code{X35}}{a numeric vector}
    \item{\code{X36}}{a numeric vector}
    \item{\code{X37}}{a numeric vector}
    \item{\code{X38}}{a numeric vector}
    \item{\code{X39}}{a numeric vector}
    \item{\code{X40}}{a numeric vector}
    \item{\code{X41}}{a numeric vector}
    \item{\code{X42}}{a numeric vector}
    \item{\code{X43}}{a numeric vector}
    \item{\code{X44}}{a numeric vector}
    \item{\code{X45}}{a numeric vector}
    \item{\code{X46}}{a numeric vector}
    \item{\code{X47}}{a numeric vector}
    \item{\code{X48}}{a numeric vector}
    \item{\code{X49}}{a numeric vector}
    \item{\code{X50}}{a numeric vector}
    \item{\code{X51}}{a numeric vector}
    \item{\code{X52}}{a numeric vector}
    \item{\code{X53}}{a numeric vector}
    \item{\code{X54}}{a numeric vector}
    \item{\code{X55}}{a numeric vector}
    \item{\code{X56}}{a numeric vector}
    \item{\code{X57}}{a numeric vector}
    \item{\code{X58}}{a numeric vector}
    \item{\code{X59}}{a numeric vector}
    \item{\code{X60}}{a numeric vector}
    \item{\code{X61}}{a numeric vector}
    \item{\code{X62}}{a numeric vector}
    \item{\code{X63}}{a numeric vector}
    \item{\code{X64}}{a numeric vector}
    \item{\code{X65}}{a numeric vector}
    \item{\code{X66}}{a numeric vector}
    \item{\code{X67}}{a numeric vector}
    \item{\code{X68}}{a numeric vector}
    \item{\code{X69}}{a numeric vector}
  }
}

\references{
Jiang L, Linton O B, Tang H, Zhang Y. Improving estimation efficiency via regression-adjustment in covariate-adaptive randomizations with imperfect compliance [J]. 2022.
}

\keyword{datasets}
