% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg.R
\name{LogisticReg}
\alias{LogisticReg}
\title{Logistic Regression Function}
\usage{
LogisticReg(x)
}
\arguments{
\item{x}{A nx1 matrix.}
}
\value{
y A nx1 matrix.
  y equals to exp(x)/(1+exp(x)) if y is not NA and 0 else.
}
\description{
Regression-adjusted Estimation of Quantile Treatment Effects under
  Covariate-Adaptive Randomizations.
}
\examples{
x <- pracma::rand(5,1)
y <- LogisticReg(x = x)

}
