% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pihat.R
\name{pihat}
\alias{pihat}
\title{Compute Estimated Pi}
\usage{
pihat(A, S)
}
\arguments{
\item{A}{A nx1 vector.}

\item{S}{A nx1 vector.}
}
\value{
A nx1 cector, each element corresponds to  the estimated pi in Jiang et al.(2022).
}
\description{
Pihat computes the estimated pi in Jiang et al.(2022) and stacks them in an nx1 vector.
}
\examples{
DGP <-FuncDGP(dgptype = 1,rndflag = 2,n = 100,g = 4,pi = 0.5)
A <- DGP[["A"]]
S <- DGP[["S"]]
pihat(A = A, S = S)

}
\references{
Jiang L, Linton O B, Tang H, et al. Improving estimation efficiency via regression-adjustment in covariate-adaptive randomizations with imperfect compliance [J]. 2022.
}
