% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinebasis.R
\name{splinebasis}
\alias{splinebasis}
\title{For each column of an input matrix, elements which are less than the median of that column are set to 0, leaving the rest of the elements unchanged}
\usage{
splinebasis(X)
}
\arguments{
\item{X}{The extra covariates, a n x K matrix. No constant included.}
}
\value{
H A n x K matrix. All elements of the X that are less than
the median of their corresponding columns are set to 0, leaving the rest unchanged.
}
\description{
For each column of an input matrix, elements which are less than the median of that column are set to 0, leaving the rest of the elements unchanged.
}
\examples{
library(pracma)
X <- rand(4,4)
H <- splinebasis(X = X)

}
