% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncDGP.R
\name{FuncDGP}
\alias{FuncDGP}
\title{Generate Data for LATE}
\usage{
FuncDGP(dgptype, rndflag, n, g, pi)
}
\arguments{
\item{dgptype}{A Scalar. 1, 2, 3 (See Jiang et al.(2022) for DGP details)}

\item{rndflag}{A Scalar. Declare the method of covariate-adaptive randomization. 1-SRS; 2-WEI; 3-BCD; 4-SBR.}

\item{n}{Sample size}

\item{g}{Number of strata. We set g=4 in the Jiang et al.(2022).}

\item{pi}{Targeted assignment probability}
}
\value{
FuncDGP returns a list containing 5 nx1 vectors named Y, X, S, A, and D.
   These five vectors correspond to the five observed vectors in Jiang et al.(2022).
   Note that vector X does not contain the constant term.
}
\description{
Generate three types of random data according to the forms in Jiang et al.(2022). FuncDGP is based on Yubo Tao's code.
}
\examples{
FuncDGP(dgptype = 1, rndflag = 1, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 1, rndflag = 2, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 1, rndflag = 3, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 1, rndflag = 4, n = 200, g = 4, pi = 0.5)

FuncDGP(dgptype = 2, rndflag = 1, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 2, rndflag = 2, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 2, rndflag = 3, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 2, rndflag = 4, n = 200, g = 4, pi = 0.5)

FuncDGP(dgptype = 3, rndflag = 1, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 3, rndflag = 2, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 3, rndflag = 3, n = 200, g = 4, pi = 0.5)
FuncDGP(dgptype = 3, rndflag = 4, n = 200, g = 4, pi = 0.5)
}
\references{
Jiang L, Linton O B, Tang H, et al. Improving estimation efficiency via regression-adjustment in covariate-adaptive randomizations with imperfect compliance [J]. 2022.
}
