% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinebasis.R
\name{splinebasis}
\alias{splinebasis}
\title{Spline The Matrix}
\usage{
splinebasis(X)
}
\arguments{
\item{X}{The extra covariate, a nxK matrix. No constant included.}
}
\value{
H A nxK matrix whose each element is bigger than its column median.
}
\description{
Splinebasis is based on Yubo's code.
   It generates polynomial spline basis in Chen (2007) p 5571.
}
\examples{
library(pracma)
X <- rand(4,4)
H <- splinebasis(X = X)

}
